+++ lib/src/calendar_controller.dart
@@ -34,42 +34,48 @@ class CalendarController {
   /// Currently visible calendar format.
   CalendarFormat get calendarFormat => _calendarFormat.value;
 
-  set events(events) {
-    _events = events;
+  /// List of currently visible days.
+  List<DateTime> get visibleDays => calendarFormat == CalendarFormat.month && !_includeInvisibleDays
+      ? _visibleDays.value.where((day) => !_isExtraDay(day)).toList()
+      : _visibleDays.value;
+
+  /// `Map` of currently visible events.
+  Map<DateTime, List> get visibleEvents {
+    if (_events == null) {
+      return {};
+    }
+
+    return Map.fromEntries(
+      _events.entries.where((entry) {
+        for (final day in visibleDays) {
+          if (_isSameDay(day, entry.key)) {
+            return true;
+          }
+        }
+
+        return false;
+      }),
+    );
   }
 
-  get events => _events;
+  /// `Map` of currently visible holidays.
+  Map<DateTime, List> get visibleHolidays {
+    if (_holidays == null) {
+      return {};
+    }
 
-  /// List of currently visible days.
-  List<DateTime> get visibleDays => _includeInvisibleDays
-      ? _visibleDays.value
-      : _visibleDays.value.where((day) => !_isExtraDay(day)).toList();
-
-  /// Map of currently visible events.
-  Map<DateTime, List> get visibleEvents => Map.fromEntries(
-        _events.entries.where((entry) {
-          for (final day in visibleDays) {
-            if (_isSameDay(day, entry.key)) {
-              return true;
-            }
+    return Map.fromEntries(
+      _holidays.entries.where((entry) {
+        for (final day in visibleDays) {
+          if (_isSameDay(day, entry.key)) {
+            return true;
           }
+        }
 
-          return false;
-        }),
-      );
-
-  /// Map of currently visible holidays.
-  Map<DateTime, List> get visibleHolidays => Map.fromEntries(
-        _holidays.entries.where((entry) {
-          for (final day in visibleDays) {
-            if (_isSameDay(day, entry.key)) {
-              return true;
-            }
-          }
-
-          return false;
-        }),
-      );
+        return false;
+      }),
+    );
+  }
 
   Map<DateTime, List> _events;
   Map<DateTime, List> _holidays;
@@ -111,7 +117,7 @@ class CalendarController {
     _dx = 0;
 
     final now = DateTime.now();
-    _focusedDay = initialDay ?? DateTime(now.year, now.month, now.day);
+    _focusedDay = initialDay ?? _normalizeDate(now);
     _selectedDay = _focusedDay;
     _calendarFormat = ValueNotifier(initialFormat);
     _visibleDays = ValueNotifier(_getVisibleDays());
@@ -147,21 +153,8 @@ class CalendarController {
   /// }
   /// ```
   void dispose() {
-    _calendarFormat.dispose();
-    _visibleDays.dispose();
-  }
-
-  String hoursToText(double hours) {
-    String displayHours = hours.toInt().toString();
-    String displayMinutes = ((hours * 60) % 60).toInt().toString();
-
-    if (displayHours.length == 1) {
-      displayHours = '0' + displayHours;
-    }
-    if (displayMinutes.length == 1) {
-      displayMinutes = '0' + displayMinutes;
-    }
-    return "$displayHours:$displayMinutes";
+    _calendarFormat?.dispose();
+    _visibleDays?.dispose();
   }
 
   /// Toggles calendar format. Same as using `FormatButton`.
@@ -199,26 +192,28 @@ class CalendarController {
     bool animate = true,
     bool runCallback = false,
   }) {
+    final normalizedDate = _normalizeDate(value);
+
     if (animate) {
-      if (value.isBefore(_getFirstDay(includeInvisible: false))) {
+      if (normalizedDate.isBefore(_getFirstDay(includeInvisible: false))) {
         _decrementPage();
-      } else if (value.isAfter(_getLastDay(includeInvisible: false))) {
+      } else if (normalizedDate.isAfter(_getLastDay(includeInvisible: false))) {
         _incrementPage();
       }
     }
 
-    _selectedDay = value;
-    _focusedDay = value;
+    _selectedDay = normalizedDate;
+    _focusedDay = normalizedDate;
     _updateVisibleDays(isProgrammatic);
 
     if (isProgrammatic && runCallback && _selectedDayCallback != null) {
-      _selectedDayCallback(value);
+      _selectedDayCallback(normalizedDate);
     }
   }
 
   /// Sets displayed month/year without changing the currently selected day.
   void setFocusedDay(DateTime value) {
-    _focusedDay = value;
+    _focusedDay = _normalizeDate(value);
     _updateVisibleDays(true);
   }
 
@@ -236,9 +231,8 @@ class CalendarController {
     return formats[id];
   }
 
-  String _getFormatButtonText() => _useNextCalendarFormat
-      ? _availableCalendarFormats[_nextFormat()]
-      : _availableCalendarFormats[_calendarFormat.value];
+  String _getFormatButtonText() =>
+      _useNextCalendarFormat ? _availableCalendarFormats[_nextFormat()] : _availableCalendarFormats[_calendarFormat.value];
 
   void _selectPrevious() {
     if (calendarFormat == CalendarFormat.month) {
@@ -280,8 +274,7 @@ class CalendarController {
       _focusedDay = _previousWeek(_focusedDay);
     } else {
       // in bottom row OR not visible
-      _focusedDay =
-          _previousWeek(_focusedDay.subtract(const Duration(days: 7)));
+      _focusedDay = _previousWeek(_focusedDay.subtract(const Duration(days: 7)));
     }
   }
 
@@ -341,29 +334,29 @@ class CalendarController {
 
   List<DateTime> _daysInMonth(DateTime month) {
     final first = _firstDayOfMonth(month);
-    final daysBefore = _startingDayOfWeek == StartingDayOfWeek.sunday
-        ? first.weekday % 7
-        : first.weekday - 1;
+    final daysBefore = _getDaysBefore(first);
     final firstToDisplay = first.subtract(Duration(days: daysBefore));
 
     final last = _lastDayOfMonth(month);
-    var daysAfter = 7 - last.weekday;
-
-    if (_startingDayOfWeek == StartingDayOfWeek.sunday) {
-      // If the last day is Sunday (7) the entire week must be rendered
-      if (daysAfter == 0) {
-        daysAfter = 7;
-      }
-    } else {
-      daysAfter++;
-    }
+    final daysAfter = _getDaysAfter(last);
 
     final lastToDisplay = last.add(Duration(days: daysAfter));
     return _daysInRange(firstToDisplay, lastToDisplay).toList();
   }
 
-  List<DateTime> daysInMonth(DateTime month) {
-    return _daysInMonth(month);
+  int _getDaysBefore(DateTime firstDay) {
+    return (firstDay.weekday + 7 - _getWeekdayNumber(_startingDayOfWeek)) % 7;
+  }
+
+  int _getDaysAfter(DateTime lastDay) {
+    int invertedStartingWeekday = 8 - _getWeekdayNumber(_startingDayOfWeek);
+
+    int daysAfter = 7 - ((lastDay.weekday + invertedStartingWeekday) % 7) + 1;
+    if (daysAfter == 8) {
+      daysAfter = 1;
+    }
+
+    return daysAfter;
   }
 
   List<DateTime> _daysInWeek(DateTime week) {
@@ -374,20 +367,16 @@ class CalendarController {
   }
 
   DateTime _firstDayOfWeek(DateTime day) {
-    day = DateTime.utc(day.year, day.month, day.day, 12);
+    day = _normalizeDate(day);
 
-    final decreaseNum = _startingDayOfWeek == StartingDayOfWeek.sunday
-        ? day.weekday % 7
-        : day.weekday - 1;
+    final decreaseNum = _getDaysBefore(day);
     return day.subtract(Duration(days: decreaseNum));
   }
 
   DateTime _lastDayOfWeek(DateTime day) {
-    day = DateTime.utc(day.year, day.month, day.day, 12);
+    day = _normalizeDate(day);
 
-    final increaseNum = _startingDayOfWeek == StartingDayOfWeek.sunday
-        ? day.weekday % 7
-        : day.weekday - 1;
+    final increaseNum = _getDaysBefore(day);
     return day.add(Duration(days: 7 - increaseNum));
   }
 
@@ -396,9 +385,7 @@ class CalendarController {
   }
 
   DateTime _lastDayOfMonth(DateTime month) {
-    final date = month.month < 12
-        ? DateTime.utc(month.year, month.month + 1, 1, 12)
-        : DateTime.utc(month.year + 1, 1, 1, 12);
+    final date = month.month < 12 ? DateTime.utc(month.year, month.month + 1, 1, 12) : DateTime.utc(month.year + 1, 1, 1, 12);
     return date.subtract(const Duration(days: 1));
   }
 
@@ -430,18 +417,26 @@ class CalendarController {
     var temp = firstDay;
 
     while (temp.isBefore(lastDay)) {
-      yield DateTime.utc(temp.year, temp.month, temp.day, 12);
+      yield _normalizeDate(temp);
       temp = temp.add(const Duration(days: 1));
     }
+  }
+
+  DateTime _normalizeDate(DateTime value) {
+    return DateTime.utc(value.year, value.month, value.day, 12);
+  }
+
+  DateTime _getEventKey(DateTime day) {
+    return visibleEvents.keys.firstWhere((it) => _isSameDay(it, day), orElse: () => null);
+  }
+
+  DateTime _getHolidayKey(DateTime day) {
+    return visibleHolidays.keys.firstWhere((it) => _isSameDay(it, day), orElse: () => null);
   }
 
   /// Returns true if `day` is currently selected.
   bool isSelected(DateTime day) {
     return _isSameDay(day, selectedDay);
-  }
-
-  bool something1() {
-    return true;
   }
 
   /// Returns true if `day` is the same day as `DateTime.now()`.
@@ -450,17 +445,11 @@ class CalendarController {
   }
 
   bool _isSameDay(DateTime dayA, DateTime dayB) {
-    return dayA.year == dayB.year &&
-        dayA.month == dayB.month &&
-        dayA.day == dayB.day;
+    return dayA.year == dayB.year && dayA.month == dayB.month && dayA.day == dayB.day;
   }
 
-  bool isSameDay(DateTime dayA, DateTime dayB) {
-    return _isSameDay(dayA, dayB);
-  }
-
-  bool _isWeekend(DateTime day) {
-    return day.weekday == DateTime.saturday || day.weekday == DateTime.sunday;
+  bool _isWeekend(DateTime day, List<int> weekendDays) {
+    return weekendDays.contains(day.weekday);
   }
 
   bool _isExtraDay(DateTime day) {